refactor(usage): drop project name from span breakdown; fold into SpanService
- Remove project_name from UsageByWorkspaceProjectUserResponse; return only project_id
- Move getSpanBreakdownPerWorkspace into SpanService and delete SpanUsageBreakdownService
- Rename SpanUsageBreakdownServiceTest back to UsageResourceTest
