+++ apps/opik-backend/src/main/java/com/comet/opik/api/resources/v1/internal/UsageResource.java
@@ -8,7 +8,6 @@ import com.comet.opik.api.UsageByWorkspaceProjectUserResponse;
 import com.comet.opik.domain.DatasetService;
 import com.comet.opik.domain.ExperimentService;
 import com.comet.opik.domain.SpanService;
-import com.comet.opik.domain.SpanUsageBreakdownService;
 import com.comet.opik.domain.TraceService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -38,7 +37,6 @@ public class UsageResource {
     private final @NonNull SpanService spanService;
     private final @NonNull ExperimentService experimentService;
     private final @NonNull DatasetService datasetService;
-    private final @NonNull SpanUsageBreakdownService spanUsageBreakdownService;
 
     @GET
     @Path("/workspace-trace-counts")
@@ -65,7 +63,7 @@ public class UsageResource {
     @Operation(operationId = "getSpansCountBreakdownForWorkspaces", summary = "Get spans count on previous day grouped by workspace, project and user", description = "Get spans count on previous day grouped by workspace, project and user", responses = {
             @ApiResponse(responseCode = "200", description = "UsageByWorkspaceProjectUserResponse resource", content = @Content(schema = @Schema(implementation = UsageByWorkspaceProjectUserResponse.class)))})
     public Response getSpansCountBreakdownForWorkspaces() {
-        return spanUsageBreakdownService.getSpanBreakdownPerWorkspace()
+        return spanService.getSpanBreakdownPerWorkspace()
                 .map(breakdownResponse -> Response.ok(breakdownResponse).build())
                 .block();
     }
